backend/requirements.txt
numpy==2.4.4
packaging==26.0
pandas==3.0.0
pillow==12.1.1
PyMuPDF==1.27.2
pytesseract==0.3.13
sentence-transformers==2.7.0
faiss-cpu==1.8.0
watchdog==6.0
pdf2image==1.16